requirements.txt
Pillow==7.2.0
python-vlc==3.0.11115
